Parsed to render refactoring
