+++ src/generation/mod.rs
@@ -1,2 +1,8 @@
+pub mod json;
 pub mod parsed_to_render;
 mod test;
+
+pub use self::{
+    json::to_json,
+    parsed_to_render::{r_to_f64, render, to_wav},
+};
